Extract shared Discord E2E test helpers into helpers module
