+++ src/discord/e2e/helpers.ts
@@ -1,0 +1,63 @@
+import fs from "node:fs";
+import os from "node:os";
+import path from "node:path";
+
+/**
+ * Generate a standardized E2E channel name using the local
+ * timestamp: `e2e-YYYY-MM-DD-t-HH-MM-SS`.
+ */
+export function e2eChannelName(): string {
+  const now = new Date();
+  const yyyy = now.getFullYear();
+  const mm = String(now.getMonth() + 1).padStart(2, "0");
+  const dd = String(now.getDate()).padStart(2, "0");
+  const hh = String(now.getHours()).padStart(2, "0");
+  const min = String(now.getMinutes()).padStart(2, "0");
+  const ss = String(now.getSeconds()).padStart(2, "0");
+  return `e2e-${yyyy}-${mm}-${dd}-t-${hh}-${min}-${ss}`;
+}
+
+export function resolveTestBotToken(): string {
+  if (process.env.DISCORD_E2E_BOT_TOKEN) {
+    return process.env.DISCORD_E2E_BOT_TOKEN;
+  }
+  const keyPath = path.join(os.homedir(), ".keys", "discord-e2e-bot-token");
+  try {
+    return fs.readFileSync(keyPath, "utf-8").trim();
+  } catch {
+    throw new Error(
+      `Discord E2E bot token not found. Set DISCORD_E2E_BOT_TOKEN or ` +
+        `create ${keyPath} with the token.`,
+    );
+  }
+}
+
+export type MessageEvent = {
+  type: "create" | "update" | "delete";
+  messageId: string;
+  content?: string;
+  timestamp: number;
+};
+
+export async function waitForBotResponse(
+  events: MessageEvent[],
+  maxWaitMs: number,
+  quietPeriodMs: number,
+): Promise<void> {
+  const startTime = Date.now();
+  let lastEventTime = startTime;
+
+  while (Date.now() - startTime < maxWaitMs) {
+    await new Promise((r) => setTimeout(r, 1000));
+
+    const latestEvent = events[events.length - 1];
+    if (latestEvent) {
+      lastEventTime = latestEvent.timestamp;
+    }
+
+    const creates = events.filter((e) => e.type === "create");
+    if (creates.length > 0 && Date.now() - lastEventTime >= quietPeriodMs) {
+      break;
+    }
+  }
+}
