+++ src/discord/e2e/mid-turn-messaging.e2e.test.ts
@@ -1,10 +1,13 @@
 import { ChannelType, Client, Events, GatewayIntentBits } from "discord.js";
 import { randomBytes } from "node:crypto";
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import { isTruthyEnvValue } from "../../infra/env.js";
+import {
+  type MessageEvent,
+  e2eChannelName,
+  resolveTestBotToken,
+  waitForBotResponse,
+} from "./helpers.js";
 
 // Gated behind LIVE=1 — these tests hit real Discord.
 const LIVE = isTruthyEnvValue(process.env.LIVE) || isTruthyEnvValue(process.env.CLAWDBOT_LIVE_TEST);
@@ -14,54 +17,6 @@ const describeLive = LIVE ? describe : describe.skip;
 const CLAW_BOT_ID = process.env.DISCORD_E2E_CLAW_BOT_ID ?? "1468764779471700133";
 // Guild where the E2E tester bot can create channels.
 const GUILD_ID = process.env.DISCORD_E2E_GUILD_ID ?? "1471323114418733261";
-
-function resolveTestBotToken(): string {
-  if (process.env.DISCORD_E2E_BOT_TOKEN) {
-    return process.env.DISCORD_E2E_BOT_TOKEN;
-  }
-  const keyPath = path.join(os.homedir(), ".keys", "discord-e2e-bot-token");
-  try {
-    return fs.readFileSync(keyPath, "utf-8").trim();
-  } catch {
-    throw new Error(
-      `Discord E2E bot token not found. Set DISCORD_E2E_BOT_TOKEN or ` +
-        `create ${keyPath} with the token.`,
-    );
-  }
-}
-
-type MessageEvent = {
-  type: "create" | "update" | "delete";
-  messageId: string;
-  content?: string;
-  timestamp: number;
-};
-
-/** Wait for the bot to finish responding in a channel. Returns when
- * at least one `create` event exists and `quietPeriodMs` elapses
- * with no new events, or `maxWaitMs` total time has elapsed. */
-async function waitForBotResponse(
-  events: MessageEvent[],
-  maxWaitMs: number,
-  quietPeriodMs: number,
-): Promise<void> {
-  const startTime = Date.now();
-  let lastEventTime = startTime;
-
-  while (Date.now() - startTime < maxWaitMs) {
-    await new Promise((r) => setTimeout(r, 1000));
-
-    const latestEvent = events[events.length - 1];
-    if (latestEvent) {
-      lastEventTime = latestEvent.timestamp;
-    }
-
-    const creates = events.filter((e) => e.type === "create");
-    if (creates.length > 0 && Date.now() - lastEventTime >= quietPeriodMs) {
-      break;
-    }
-  }
-}
 
 /** Wait until we see at least one bot message, indicating the run
  * has started. Gives up after `maxWaitMs`. */
@@ -79,8 +34,6 @@ async function waitForFirstBotMessage(events: MessageEvent[], maxWaitMs: number)
 describeLive("Discord mid-turn messaging (steer mode)", () => {
   let client: Client;
   const nonce = randomBytes(4).toString("hex");
-  const now = new Date();
-  const today = `${now.getFullYear()}-${String(now.getMonth() + 1).padStart(2, "0")}-${String(now.getDate()).padStart(2, "0")}`;
 
   let channelId: string;
   const events: MessageEvent[] = [];
@@ -108,9 +61,8 @@ describeLive("Discord mid-turn messaging (steer mode)", () => {
 
     // Create an isolated channel for this test.
     const guild = await client.guilds.fetch(GUILD_ID);
-    const channelName = `e2e-${today}-mid-turn-${nonce}`;
     const channel = await guild.channels.create({
-      name: channelName,
+      name: e2eChannelName(),
       type: ChannelType.GuildText,
       topic: `E2E mid-turn messaging test (auto-created, safe to delete)`,
     });
