+++ src/discord/e2e/multi-tool-feedback.e2e.test.ts
@@ -5,6 +5,12 @@ import os from "node:os";
 import path from "node:path";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import { isTruthyEnvValue } from "../../infra/env.js";
+import {
+  type MessageEvent,
+  e2eChannelName,
+  resolveTestBotToken,
+  waitForBotResponse,
+} from "./helpers.js";
 
 // Gated behind LIVE=1 — these tests hit real Discord.
 const LIVE = isTruthyEnvValue(process.env.LIVE) || isTruthyEnvValue(process.env.CLAWDBOT_LIVE_TEST);
@@ -14,28 +20,6 @@ const describeLive = LIVE ? describe : describe.skip;
 const CLAW_BOT_ID = process.env.DISCORD_E2E_CLAW_BOT_ID ?? "1468764779471700133";
 // Guild where the E2E tester bot can create channels.
 const GUILD_ID = process.env.DISCORD_E2E_GUILD_ID ?? "1471323114418733261";
-
-function resolveTestBotToken(): string {
-  if (process.env.DISCORD_E2E_BOT_TOKEN) {
-    return process.env.DISCORD_E2E_BOT_TOKEN;
-  }
-  const keyPath = path.join(os.homedir(), ".keys", "discord-e2e-bot-token");
-  try {
-    return fs.readFileSync(keyPath, "utf-8").trim();
-  } catch {
-    throw new Error(
-      `Discord E2E bot token not found. Set DISCORD_E2E_BOT_TOKEN or ` +
-        `create ${keyPath} with the token.`,
-    );
-  }
-}
-
-type MessageEvent = {
-  type: "create" | "update" | "delete";
-  messageId: string;
-  content?: string;
-  timestamp: number;
-};
 
 // Known tool feedback patterns in the rich format (*ToolName*) and
 // the older italic format (*Verbing ...*).
@@ -60,32 +44,6 @@ const TOOL_FEEDBACK_PATTERNS = [
   "*Fetching",
   "*Globbing",
 ];
-
-/** Wait for the bot to finish responding in a channel. Returns when
- * at least one `create` event exists and `quietPeriodMs` elapses
- * with no new events, or `maxWaitMs` total time has elapsed. */
-async function waitForBotResponse(
-  events: MessageEvent[],
-  maxWaitMs: number,
-  quietPeriodMs: number,
-): Promise<void> {
-  const startTime = Date.now();
-  let lastEventTime = startTime;
-
-  while (Date.now() - startTime < maxWaitMs) {
-    await new Promise((r) => setTimeout(r, 1000));
-
-    const latestEvent = events[events.length - 1];
-    if (latestEvent) {
-      lastEventTime = latestEvent.timestamp;
-    }
-
-    const creates = events.filter((e) => e.type === "create");
-    if (creates.length > 0 && Date.now() - lastEventTime >= quietPeriodMs) {
-      break;
-    }
-  }
-}
 
 /** Check that no tool result content is duplicated excessively.
  * It is acceptable for content to appear in a tool result block AND
@@ -155,8 +113,6 @@ type SuiteName = (typeof TEST_SUITES)[number];
 describeLive("Discord multi-tool feedback display", () => {
   let client: Client;
   const nonce = randomBytes(4).toString("hex");
-  const now = new Date();
-  const today = `${now.getFullYear()}-${String(now.getMonth() + 1).padStart(2, "0")}-${String(now.getDate()).padStart(2, "0")}`;
   const probePath = path.join(os.tmpdir(), `e2e-probe-${nonce}.txt`);
 
   // Per-suite state: channel ID and captured events.
@@ -189,9 +145,8 @@ describeLive("Discord multi-tool feedback display", () => {
     // Create one channel per test suite.
     const guild = await client.guilds.fetch(GUILD_ID);
     for (const suite of TEST_SUITES) {
-      const channelName = `e2e-${today}-${suite}-${nonce}`;
       const channel = await guild.channels.create({
-        name: channelName,
+        name: e2eChannelName(),
         type: ChannelType.GuildText,
         topic: `E2E multi-tool feedback test: ${suite} (auto-created, safe to delete)`,
       });
